refactor: extract BitAccumulator, use it in codec, rename test file
Extract BitAccumulator from Environment.kt into its own file so it can
be used by both the simulator and the grpc codec. Rewrite packPacketOut
and stripPacketInHeader to use BitAccumulator instead of hand-rolled
bit-shifting loops.

Rename InlineP4CompilerTest → BitLevelSerializationTest to reflect its
actual scope (bit-level packet serialization, not the inline compiler).
