+++ grpc/BUILD.bazel
@@ -213,10 +213,10 @@ kt_jvm_test(
 )
 
 kt_jvm_test(
-    name = "InlineP4CompilerTest",
+    name = "BitLevelSerializationTest",
     srcs = [
+        "BitLevelSerializationTest.kt",
         "FourwardTestHarness.kt",
-        "InlineP4CompilerTest.kt",
     ],
     data = [
         "//p4c_backend:p4c-4ward",
@@ -227,7 +227,7 @@ kt_jvm_test(
         "-Dp4c_4ward=$(rlocationpath //p4c_backend:p4c-4ward)",
         "-Dp4include=$(rlocationpath @p4c//p4include:core.p4)",
     ],
-    test_class = "fourward.grpc.InlineP4CompilerTest",
+    test_class = "fourward.grpc.BitLevelSerializationTest",
     deps = [
         ":dataplane_java_proto",
         ":dataplane_kt_grpc",
